[test] 2단계 - 문자열 덧셈 계산기
추가 코드리뷰 반영 및 테스트 코드 수정

1. 패키지명 영문수정
2. 어플리케이션 코드에서 지역변수 사용시점 선언
3. 불변객체 상수선언
4. 테스트코드 인스턴스 변수 최상단
5. NullAndEmptySource 활용
