+++ src/test/java/step1_TestPractice/StringClassTest.java
@@ -1,3 +1,5 @@
+package step1_TestPractice;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -5,10 +7,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 public class StringClassTest {
-
+    private final String require2NumberData = "(1,2)";
+    private final String require2abcData = "abc";
     private final String numbersData = "1,2";
     private final String numberData = "1";
-    private String[] array;
 
     @Test
     @DisplayName(value = "1,2 split으로 분리여부 확인 테스트")
@@ -22,15 +24,11 @@ public class StringClassTest {
         assertThat(numberData.split(",")).containsExactlyInAnyOrder("1");
     }
 
-    private final String require2NumberData = "(1,2)";
-
     @Test
     @DisplayName(value = "(1,2) subString사용해서 1,2만 반환 테스트")
     public void shouldReturnNumberTest() {
         assertThat(require2NumberData.substring(1, 4)).isEqualTo("1,2");
     }
-
-    private final String require2abcData = "abc";
 
     @Test
     @DisplayName(value = "abc에서 특정위치 문자 가져오는 테스트")
